add try finally blocks
