+++ test/integration_new/object/file_itest.py
@@ -11,8 +11,6 @@ from test.integration_new.context_managers.box_retention_policy import BoxRetent
 from test.integration_new import util
 from test.integration_new.context_managers.box_test_file import BoxTestFile
 from test.integration_new.context_managers.box_test_folder import BoxTestFolder
-from test.integration_new.context_managers.box_sign_request import BoxTestSignRequest
-from test.integration_new import CLIENT
 
 
 FILE_TESTS_DIRECTORY_NAME = 'file-integration-tests'
@@ -103,17 +101,19 @@ def test_lock_and_unlock(parent_folder, small_file_path, small_file_v2_path, oth
     with BoxTestFile(parent_folder=parent_folder, file_path=small_file_path) as file:
         file.collaborate(accessible_by=other_user, role='editor')
 
-        file.lock(prevent_download=True)
+        try:
+            file.lock(prevent_download=True)
 
-        assert file.get(fields=('lock',)).lock
+            assert file.get(fields=('lock',)).lock
 
-        with pytest.raises(BoxAPIException):
-            other_client.file(file.object_id).update_contents(small_file_v2_path)
+            with pytest.raises(BoxAPIException):
+                other_client.file(file.object_id).update_contents(small_file_v2_path)
 
-        with pytest.raises(BoxAPIException):
-            other_client.file(file.object_id).download_to(io.BytesIO())
+            with pytest.raises(BoxAPIException):
+                other_client.file(file.object_id).download_to(io.BytesIO())
 
-        file.unlock()
+        finally:
+            file.unlock()
 
         assert file.get(fields=('lock',)).lock is None
 
@@ -241,55 +241,12 @@ def test_copy(test_file, parent_folder):
     copied_file_name = util.random_name()
     copied_file = test_file.copy(parent_folder=parent_folder, name=copied_file_name)
 
-    assert copied_file.id != test_file
-    assert copied_file.name == copied_file_name
-    assert test_file.content() == copied_file.content()
-
-    util.permanently_delete(copied_file)
-
-
-def test_webhook_sign_request(parent_folder, small_file_path):
-    with BoxTestFile(parent_folder=parent_folder, file_path=small_file_path) as test_file:
-        file = {
-            'id': test_file.object_id,
-            'type': test_file.object_type,
-        }
-        files = [file]
-        signer = {
-            'email': 'signer@mail.com',
-        }
-        signers = [signer]
-
-        with BoxTestSignRequest(files=files, signers=signers, parent_folder_id=parent_folder.id) as sign_request:
-            webhook_url = 'https://example.com/webhook'
-            new_webhook_url = 'https://example.com/new-webhook'
-
-            assert len(sign_request.sign_files['files']) == 1
-
-            sign_file = sign_request.sign_files['files'][0]
-            webhook_file = CLIENT.file(file_id=sign_file['id'])
-            webhook = CLIENT.create_webhook(
-                webhook_file,
-                ['SIGN_REQUEST.COMPLETED', 'SIGN_REQUEST.DECLINED', 'SIGN_REQUEST.EXPIRED'],
-                webhook_url
-            )
-
-            assert webhook.id
-            assert len(webhook.triggers) == 3
-            assert webhook.address == webhook_url
-
-            update_data = {
-                'address': new_webhook_url,
-                'triggers': ['SIGN_REQUEST.COMPLETED']
-            }
-            webhook = webhook.update_info(data=update_data)
-
-            assert webhook.address == new_webhook_url
-            assert len(webhook.triggers) == 1
-
-            webhook.delete()
-            with pytest.raises(BoxAPIException):
-                webhook.get()
+    try:
+        assert copied_file.id != test_file
+        assert copied_file.name == copied_file_name
+        assert test_file.content() == copied_file.content()
+    finally:
+        util.permanently_delete(copied_file)
 
 
 def test_set_disposition_at(parent_folder, small_file_path):
